[add] support undo auto fill
