+++ gradio_web/controllers/utils_controller.py
@@ -5,8 +5,8 @@ import time
 from PIL import Image
 
 from modules.utils.check_server_status import  check_chat_api_chat, check_sd_api_img2img, check_sd_api_loras, check_sd_api_txt2img
-from modules.utils.colors import generate_mask_from_rgb
-from modules.utils.img_segment import get_mask_by_blackpoints,replace_black_pixels
+from modules.utils.colors import generate_mask_from_black
+from modules.utils.img_segment import auto_fill_by_blackpoints,replace_black_pixels
 
 last_editor = []
 
@@ -36,19 +36,19 @@ def check_status_process():
 
 def submit_mask_process(painted):
     new_editor = gr.ImageEditor(value={"background":painted["composite"],"layers":[],"composite":None}, label='Edit', type='pil', interactive=True)
-    return generate_mask_from_rgb(painted['composite']), new_editor
+    return generate_mask_from_black(painted['composite']), new_editor
 
 def send_to_editor_process(base_img):
     new_editor = gr.ImageEditor(value={"background":replace_black_pixels(base_img),"layers":[],"composite":None}, label='Edit', type='pil', interactive=True)
     global last_editor
-    last_editor.append({"background":replace_black_pixels(base_img),"layers":[],"composite":None})
+    last_editor = [{"background":replace_black_pixels(base_img),"layers":[],"composite":None}]
     return new_editor
 
-def auto_mask_process(painted):
+def auto_mask_process(painted, init_img):
     global last_editor
     last_editor.append(painted)
     composite_img = painted["composite"]
-    auto_mask_img = get_mask_by_blackpoints(composite_img)
+    auto_mask_img = auto_fill_by_blackpoints(composite_img, init_img)
     print(auto_mask_img.mode)
     new_editor = gr.ImageEditor(value={"background":auto_mask_img,"layers":[],"composite":None}, label='Edit', type='pil', interactive=True)
     return new_editor
