+++ gradio_web/modules/api/pics_api.py
@@ -1,11 +1,13 @@
 import base64
+import functools
 from io import BytesIO
 import json
 import sys
 import gradio as gr
 from PIL import Image
 import requests
-from modules.utils.scripts_gen import form_alwayson_scripts_from_kwargv,form_default_paras_from_template
+from modules.utils.scripts_gen import form_alwayson_scripts_from_kwargv
+from modules.utils.imaga_paras_gen import form_default_paras_from_template
 
 from const import *
 
@@ -63,150 +65,8 @@ def get_loras():
     except:
         return [],f'[SD] Get loras failed'
 
-def form_post_txt2img_paras(query: str, loras:list[str]=[], **kwargv):
-    """
-    ## Form paras for post_txt2img
 
-    ## Argvs:
-    ```
-        query(str)
-        loras(list[str]): list of lora name
-
-        init_img_str(str|None)
-        alwayson_scripts(dict|None)
-        template(str|None): name of img gen template
-        prompt(str|None)
-        negative_prompt(str|None)
-        denoising_strength(float|None)
-        sampler_index(str|None)
-        seed(int|None)
-        steps(int|None)
-        width(int|None)
-        height(int|None)
-        cfg_scale(int|None)
-        alwayson_scripts(str|None)
-    ```
-    ## Return:
-    ```
-        image_str(str)
-        err_string(str|None)
-    ```
-        
-    """
-    # Load loras to query
-    for lora_name in loras:
-        query += f" <lora:{lora_name}:{1/len(loras)}> "
-
-    # ------------------------------------------------------
-    # Load always on scripts
-    alwayson_scripts, e = form_alwayson_scripts_from_kwargv(**kwargv)
-    if e != None:
-        return None, f"[SD] alwayson_scripts failed, {e}"
-    
-    # ------------------------------------------------------
-    # Set paras
-    paras = {
-        "prompt": query + kwargv.get("prompt", ",(recherche details) ,(ultra details),8k resolution,excellent quality,beautiful cinematic lighting,engaging atmosphere"), # TODO : remove this hard code
-        "negative_prompt": kwargv.get("negative_prompt", "(worst quality, low quality, cgi, bad eye, worst eye, illustration, cartoon),deformed,distorted,disfigured,poorly drawn,bad anatomy,wrong anatomy"), # TODO : remove this hard code
-        "denoising_strength": kwargv.get("denoising_strength", 0.2),
-        "sampler_index": kwargv.get("sampler_index", "DPM++ 2M Karras"),
-        "seed": kwargv.get("seed", -1),
-        "steps": kwargv.get("steps",40),
-        "width": kwargv.get("width",512),
-        "height": kwargv.get("height",512),
-        "cfg_scale": kwargv.get("cfg_scale",5),
-        "alwayson_scripts": alwayson_scripts,
-    }
-
-    return paras, None
-    
-def form_post_img2img_paras(init_img_str:str ,query: str ,loras:list[str]=[], **kwargv):
-    """
-    ## Form paras for post_img2img
-
-    ## Argvs:
-    ```
-        init_img_str(str)
-        query(str)
-        loras(list[str]): list of lora name
-
-        alwayson_scripts(dict|None)
-        template(str|None): name of img gen template
-        prompt(str|None)
-        negative_prompt(str|None)
-        denoising_strength(float|None)
-        sampler_index(str|None)
-        seed(int|None)
-        steps(int|None)
-        width(int|None)
-        height(int|None)
-        cfg_scale(int|None)
-        alwayson_scripts(str|None)
-
-        mask_img_str(str|None): mask image string
-    ```
-    ## Return:
-    ```
-        paras(dict)
-        err_string(str|None)
-    ```
-        
-    """
-    # ------------------------------------------------------
-    # Check if init_img_str valid
-    if init_img_str == None:
-        return None, f"No init image"
-    try:
-        image = Image.open(BytesIO(base64.b64decode(init_img_str)))
-    except:
-        return None, f"Invalid init image"
-    
-    # ------------------------------------------------------
-    # Load loras to query
-    for lora_name in loras:
-        query += f" <lora:{lora_name}:{1/len(loras)}> "
-
-    # ------------------------------------------------------
-    # Load alwayson_scripts
-    alwayson_scripts, e = form_alwayson_scripts_from_kwargv(init_img_str=init_img_str ,**kwargv)
-    if e != None:
-        return None, f"{e}"
-    
-    # ------------------------------------------------------
-    # Load template paras
-    template_paras, e = form_default_paras_from_template(**kwargv)
-    if e != None:
-        return None, f"{e}"
-    
-    # ------------------------------------------------------
-    # Set paras
-    # For each para, if it is not given in kwargv, it will search template. If template is not available, it will set to default
-    default_paras = IMG2IMG_DEFAULT_PARAS
-    paras = {
-        "init_images": [init_img_str],
-        "prompt": kwargv.get("prompt", template_paras.get("prompt", None)),
-        "negative_prompt": kwargv.get("negative_prompt", template_paras.get("negative_prompt", None)),
-        "denoising_strength": kwargv.get("denoising_strength", template_paras.get("denoising_strength", None)),
-        "sampler_index": kwargv.get("sampler_index", template_paras.get("sampler_index", None)),
-        "seed": kwargv.get("seed", template_paras.get("seed", None)),
-        "steps": kwargv.get("steps", template_paras.get("steps", None)),
-        "width": kwargv.get("width", template_paras.get("width", None)),
-        "height": kwargv.get("height", template_paras.get("height", None)),
-        "cfg_scale": kwargv.get("cfg_scale", template_paras.get("cfg_scale", None)),
-        "alwayson_scripts": alwayson_scripts,
-
-        "mask_image": kwargv.get("mask_img_str", template_paras.get("mask_image",None)),
-        "source": kwargv.get("source", template_paras.get("source",None))
-    }
-
-    for key in IMG2IMG_DEFAULT_PARAS.keys():
-        paras[key] = paras[key] if paras[key] != None else default_paras[key]
-        if key == "prompt":
-            paras[key] = query + paras[key]
-    
-    # ------------------------------------------------------
-    return paras, None
-
+@functools.lru_cache
 def post_hgface_img_segment(image: str):
     # ------------------------------------------------------
     # Begin check route
@@ -219,11 +79,10 @@ def post_hgface_img_segment(image: str):
     response = requests.post(SERVER_URL+route, data=json.dumps({"image": image}))
     print(response)
     try:
-        return response.json(),None
+        err = response.json().get("error", None)
+        if err != None:
+            return None, err
+        return response.json(), None
     except:
-        try:
-            print(response)
-            return None,response.json()["error"] if response.json()["error"] is not None else "Post Error"
-        except:
-            print(response)
-            return None,"Post Error"
+        print(response)
+        return None, f"Unknown err, {response}"
