+++ gradio_web/modules/utils/check_server_status.py
@@ -5,7 +5,8 @@ import threading
 import time
 
 from modules.api.chat_api import chat
-from modules.api.pics_api import get_loras,post_img2img,post_txt2img,form_post_txt2img_paras,form_post_img2img_paras
+from modules.api.pics_api import get_loras,post_img2img,post_txt2img
+from modules.utils.imaga_paras_gen import form_post_txt2img_paras,form_post_img2img_paras
 
 from const import *
