+++ gradio_web/modules/utils/img_segment.py
@@ -7,6 +7,7 @@ from PIL import Image
 import gradio as gr
 
 from modules.api.pics_api import post_hgface_img_segment
+from modules.utils.image_io import trans_image_to_str
 
 def replace_black_pixels(image: Image):
     if image == None:
@@ -25,21 +26,16 @@ def replace_black_pixels(image: Image):
     return result_image
 
 
-def process_images_by_blackpoints(original_image: Image, mask_images: dict):
+def auto_fill_black(original_image: Image, mask_images: dict):
     # 将原始图像转换为NumPy数组
     original_array = np.array(original_image)
-
     # 获取所有L模式的图像数组
     l_arrays = [np.array(mask_images[key]) for key in mask_images.keys()]
-
     # 找到原始图像中RGB为全0的像素点
     zero_pixels = np.all(original_array[..., :3] == [0, 0, 0], axis=-1)
-    
-
     # 对于每个L模式的图像数组，在原始图像中找到对应位置为全黑的像素，并将其置为0
     for layer in l_arrays:
         black_pixels = np.where(zero_pixels & (layer == 255))
-        print(black_pixels[0].size)
         if black_pixels[0].size > 10:
             try:
                 original_array[(layer == 255)] = [ 0, 0, 0, 255]
@@ -50,30 +46,30 @@ def process_images_by_blackpoints(original_image: Image, mask_images: dict):
     # 返回处理后的图像
     return result_image
 
-def get_mask_by_blackpoints(image: Image):
-    image_bytesio = io.BytesIO()
-    if image == None:
-        gr.Warning("No image provided.")
-        return None
-    image.save(image_bytesio, format="PNG")
-    img_bytes = image_bytesio.getvalue()
-    img_str = base64.b64encode(img_bytes).decode('utf-8')
-    response_json, err = post_hgface_img_segment(img_str)
-    
+def auto_fill_by_blackpoints(image: Image,init_img: Image):    
+    init_img_str = trans_image_to_str(init_img)
+    # Post huggingface models and check
+    response_json, err = post_hgface_img_segment(init_img_str)
     if err != None:
+        # clear cache
+        post_hgface_img_segment.cache_clear()
         print(err)
-        gr.Warning(err)
+        gr.Warning(err+". Cache cleared")
         return image
-    
-    labels_and_scores = [(image_package['label'],image_package['score']) for image_package in response_json["image_packages"]]
-    print(labels_and_scores)
-    
+    try:
+        labels_and_scores = [(image_package['label'],image_package['score']) for image_package in response_json["image_packages"]]
+        print(labels_and_scores)
+    except:
+        print(response_json)
+        gr.Warning(response_json.get("error", f"Unknown err: {response_json}"))
+        return image
+    # Get all mask images
     mask_images={}
     for image_package in response_json["image_packages"]:
         mask_image = Image.open(BytesIO(base64.b64decode(image_package['mask'])))
         mask_images[image_package['label']] = mask_image
     
-    result_image = process_images_by_blackpoints(image, mask_images)
+    result_image = auto_fill_black(image, mask_images)
     return result_image
 
 def process_images_by_keywords(image: Image, mask_images: dict, keys_words: list[str]):
