+++ gradio_web/test_utils.py
@@ -1,7 +1,7 @@
 import unittest
 
 from modules.utils.scripts_gen import form_alwayson_scripts_from_templates
-from modules.api.pics_api import form_post_img2img_paras,form_post_txt2img_paras
+from modules.utils.imaga_paras_gen import form_post_img2img_paras,form_post_txt2img_paras
 
 class TestSD(unittest.TestCase):
     def test_form_post_img2img_paras(self):
@@ -28,7 +28,5 @@ class TestSD(unittest.TestCase):
             init_img_str = f.read()
         alwayson_srcipts = form_alwayson_scripts_from_templates(template = "beauty",init_img_str = init_img_str)
         self.assertEqual(alwayson_srcipts[0]["controlnet"]["args"][0]["input_image"], init_img_str, "Init_img_str not applied")
-
-
 
 unittest.main()
